wktesthunter: by default only search the last 4000 revisions.
* As number of available results are growing with time, the search
can take quite a lot of time for old tests. And most of the times
only the last results are really useful. So, by default limit the
depth of the search to the last 4000 results (if the test was added
before than that). Add also some options to control this behaviour.