Adapt fetch.sh script to download json results with a random file name
* Until now the full_results.json was stored in the server on a
directory that contained the revision number. But later changes
on WebKit caused this directory to instead contain a different
number: the canonical link number@main

* But wktesthunter relies on the files stored on disk to contain
the revision number as part of the filename, so we need to find
a way of getting the revision number before storing the file on
disk.

* This patch changes the code of the fetch.sh script to get the
revision number from inside the json and then renames the json
following the previous schema.

* Since now we can't know the revision number before downloading
the json we can't use the revision number as part of the cache
of already downloaded files. We instead store the whole directory
name (which is unique) as the ID of the cache.
So, purge also all the caches files that will be regenerated later.

Related: https://github.com/clopez/webkit-testhunter/pull/9